refactor tools/internals + utils
