+++ go/tools/pkg/cilium/cilium_test.go
