+++ go/tools/pkg/utils/datetime_test.go
@@ -1,4 +1,4 @@
-package datetime
+package utils
 
 import (
 	"context"
