+++ helm/kagent/templates/deployment.yaml
@@ -52,12 +52,12 @@ spec:
             {{- toYaml . | nindent 12 }}
           {{- end }}
           ports:
-            - name: controller
+            - name: http-controller
               containerPort: {{ .Values.service.ports.controller.targetPort }}
               protocol: TCP
           readinessProbe:
             tcpSocket:
-              port: controller
+              port: http-controller
             initialDelaySeconds: 15
             periodSeconds: 15
         - name: app
@@ -80,7 +80,7 @@ spec:
             {{- toYaml . | nindent 12 }}
           {{- end }}
           ports:
-            - name: app
+            - name: http-app
               containerPort: {{ .Values.service.ports.app.targetPort }}
               protocol: TCP
           resources:
@@ -88,7 +88,7 @@ spec:
           readinessProbe:
             httpGet:
               path: /api/version
-              port: app
+              port: http-app
             initialDelaySeconds: 15
             periodSeconds: 15
         - name: ui
@@ -103,7 +103,7 @@ spec:
             {{- toYaml . | nindent 12 }}
           {{- end }}
           ports:
-            - name: ui
+            - name: http-ui
               containerPort: {{ .Values.service.ports.ui.targetPort }}
               protocol: TCP
           resources:
@@ -111,7 +111,7 @@ spec:
           readinessProbe:
             httpGet:
               path: /
-              port: ui
+              port: http-ui
             initialDelaySeconds: 15
             periodSeconds: 15
         - name: tools
@@ -141,12 +141,12 @@ spec:
             - name: OTEL_EXPORTER_OTLP_TRACES_INSECURE
               value: {{ .Values.otel.tracing.exporter.otlp.insecure | quote }}
           ports:
-            - name: tools
+            - name: http-tools
               containerPort: {{ .Values.service.ports.tools.targetPort }}
               protocol: TCP
           readinessProbe:
             tcpSocket:
-              port: tools
+              port: http-tools
             initialDelaySeconds: 15
             periodSeconds: 15
         {{ if and .Values.mcp.doc2vec.enabled (eq .Values.providers.default "openAI") }}
@@ -183,7 +183,7 @@ spec:
             - name: OTEL_EXPORTER_OTLP_TRACES_INSECURE
               value: {{ .Values.otel.tracing.exporter.otlp.insecure | quote }}
           ports:
-            - name: querydoc
+            - name: http-querydoc
               containerPort: {{ .Values.service.ports.querydoc.targetPort }}
               protocol: TCP
         {{ end }}
