+++ helm/kagent/values.yaml
@@ -89,7 +89,7 @@ mcp:
     image:
       registry: ghcr.io
       repository: kagent-dev/doc2vec/mcp
-      tag: 1.1.5
+      tag: 1.1.7
       pullPolicy: IfNotPresent
     resources:
       requests:
